+++ test/Fees.t.sol
@@ -162,15 +162,15 @@ contract FeesTest is Test, BaseTestHelper {
         uint256 rail1FirstExpectedAmount = RAIL1_RATE * 5; // 5 blocks * 5 ether
         
         // Settle rail1 (token1)
-        settlementHelper.settleRailAndVerify(rail1Id, block.number, rail1FirstExpectedAmount, block.number);
+        settlementHelper.settleRailAndVerify(rail1Id, uint64(block.number), rail1FirstExpectedAmount, uint64(block.number));
         
         // Settle rail2 (token2)
         vm.prank(USER1);
-        (uint256 settledAmount2,,,,, ) = payments.settleRail(rail2Id, block.number);
+        (uint256 settledAmount2,,,,, ) = payments.settleRail(rail2Id, uint64(block.number));
         
         // Settle rail3 (token3)
         vm.prank(USER1);
-        (uint256 settledAmount3,,,,, ) = payments.settleRail(rail3Id, block.number);
+        (uint256 settledAmount3,,,,, ) = payments.settleRail(rail3Id, uint64(block.number));
         
         // Calculate expected fees based on actual settled amounts (0.1% of settled amounts)
         uint256 rail1FirstFee = (rail1FirstExpectedAmount * payments.PAYMENT_FEE_BPS()) / payments.COMMISSION_MAX_BPS();
@@ -202,15 +202,15 @@ contract FeesTest is Test, BaseTestHelper {
         uint256 rail1SecondExpectedAmount = RAIL1_RATE * 7; // 7 blocks * 5 ether
         
         // Settle rail1 (token1) again
-        settlementHelper.settleRailAndVerify(rail1Id, block.number, rail1SecondExpectedAmount, block.number);
+        settlementHelper.settleRailAndVerify(rail1Id, uint64(block.number), rail1SecondExpectedAmount, uint64(block.number));
         
         // Settle rail2 (token2) again
         vm.prank(USER1);
-        (uint256 secondSettledAmount2,,,,, ) = payments.settleRail(rail2Id, block.number);
+        (uint256 secondSettledAmount2,,,,, ) = payments.settleRail(rail2Id, uint64(block.number));
         
         // Settle rail3 (token3) again
         vm.prank(USER1);
-        (uint256 secondSettledAmount3,,,,, ) = payments.settleRail(rail3Id, block.number);
+        (uint256 secondSettledAmount3,,,,, ) = payments.settleRail(rail3Id, uint64(block.number));
         
         // Calculate expected fees for second round - use actual settled amounts
         uint256 rail1SecondFee = (rail1SecondExpectedAmount * payments.PAYMENT_FEE_BPS()) / payments.COMMISSION_MAX_BPS();
